+++ library/docs/changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Navigation from supervised feature to supervisor.
+- Amount of offscreen feature sections in inspector can be controlled with `LaboratoryActivity.Configuration.OffscreenSectionsBehavior`. Default is unlimited.
 
 ### Changed
 - Upgrade to Kotlin `1.4.32`.